Remove DisplacementGradientFunction wrapper class, extract Incremental classes to separate module
